refactor: Вынести ExceptionHandler в отдельный класс style: убрать закомментированный код
