+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -2,8 +2,6 @@ package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
@@ -17,70 +15,52 @@ import java.util.List;
 public class UserController {
 
 
-	private final UserService userService;
+    private final UserService userService;
 
-	@Autowired
-	public UserController(UserService userService) {
-		this.userService = userService;
-	}
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
 
-	@PutMapping
-	public User updateUser(@Valid @RequestBody User user) {
-		return userService.updateUser(user);
-	}
+    @PutMapping
+    public User updateUser(@Valid @RequestBody User user) {
+        return userService.updateUser(user);
+    }
 
-	@PostMapping
-	public User addUser(@Valid @RequestBody User user) {
-		return userService.addUser(user);
-	}
+    @PostMapping
+    public User addUser(@Valid @RequestBody User user) {
+        return userService.addUser(user);
+    }
 
-	@GetMapping
-	public List<User> getAllUsers() {
-		log.debug("get all users");
-		return userService.getAllUsers();
-	}
+    @GetMapping
+    public List<User> getAllUsers() {
+        log.debug("get all users");
+        return userService.getAllUsers();
+    }
 
-	@GetMapping("/{userId}")
-	public User getUserById(@PathVariable Integer userId) {
-		return userService.getUserById(userId);
-	}
+    @GetMapping("/{userId}")
+    public User getUserById(@PathVariable Integer userId) {
+        return userService.getUserById(userId);
+    }
 
-	@PutMapping("/{id}/friends/{friendId}")
-	public User addFriend(@PathVariable Integer id, @PathVariable Integer friendId) {
-		return userService.addFriend(id, friendId, true);
-	}
+    @PutMapping("/{id}/friends/{friendId}")
+    public User addFriend(@PathVariable Integer id, @PathVariable Integer friendId) {
+        return userService.addFriend(id, friendId, true);
+    }
 
-	@DeleteMapping("/{id}/friends/{friendId}")
-	public User deleteFriend(@PathVariable Integer id, @PathVariable Integer friendId) {
-		return userService.addFriend(id, friendId, false);
-	}
+    @DeleteMapping("/{id}/friends/{friendId}")
+    public User deleteFriend(@PathVariable Integer id, @PathVariable Integer friendId) {
+        return userService.addFriend(id, friendId, false);
+    }
 
-	@GetMapping("/{userId}/friends")
-	public List<User> getAllFriends(@PathVariable int userId) {
-		log.debug("get all friends for user {}", userId);
-		return userService.getAllFriends(userId);
-	}
+    @GetMapping("/{userId}/friends")
+    public List<User> getAllFriends(@PathVariable int userId) {
+        log.debug("get all friends for user {}", userId);
+        return userService.getAllFriends(userId);
+    }
 
-	@GetMapping("/{id}/friends/common/{otherId}")
-	public List<User> getIntersectionFriends(@PathVariable Integer id, @PathVariable Integer otherId) {
-		return userService.getIntersectionFriends(id, otherId);
-	}
-
-	@ExceptionHandler
-	@ResponseStatus(HttpStatus.BAD_REQUEST)
-	public ErrorResponse handle(final MethodArgumentNotValidException e) {
-		return new ErrorResponse(
-				"Data validation error", e.getMessage()
-		);
-	}
-
-	@ExceptionHandler
-	@ResponseStatus(HttpStatus.NOT_FOUND)
-	public ErrorResponse handle(final RuntimeException e) {
-		return new ErrorResponse(
-				"No data found", e.getMessage()
-		);
-	}
-
-
+    @GetMapping("/{id}/friends/common/{otherId}")
+    public List<User> getIntersectionFriends(@PathVariable Integer id, @PathVariable Integer otherId) {
+        return userService.getIntersectionFriends(id, otherId);
+    }
 }
