+++ src/main/resources/schema.sql
@@ -10,14 +10,14 @@ DROP TABLE IF EXISTS users;
 
 CREATE TABLE IF NOT EXISTS films(
 id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-name VARCHAR,
-description VARCHAR,
+name VARCHAR(50),
+description VARCHAR(200),
 release_date DATE,
 duration INT);
 
 CREATE TABLE IF NOT EXISTS genres(
 genre_id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-genre_name VARCHAR);
+genre_name VARCHAR(50));
 
 CREATE TABLE IF NOT EXISTS films_genres(
 id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
@@ -26,7 +26,7 @@ genre_id INTEGER REFERENCES genres(genre_id) ON DELETE CASCADE);
 
 CREATE TABLE IF NOT EXISTS mpa(
 id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-mpa_name VARCHAR);
+mpa_name VARCHAR(10));
 
 CREATE TABLE IF NOT EXISTS films_mpa(
 film_id INTEGER PRIMARY KEY,
@@ -35,9 +35,9 @@ mpa_id INTEGER REFERENCES mpa(id));
 
 CREATE TABLE IF NOT EXISTS users(
 id INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-email VARCHAR,
-login VARCHAR NOT NULL,
-name VARCHAR,
+email VARCHAR(200),
+login VARCHAR(200) NOT NULL,
+name VARCHAR(200),
 birthday DATE);
 
 CREATE TABLE IF NOT EXISTS likes_film(
@@ -52,6 +52,3 @@ user1_id INTEGER REFERENCES users(id),
 user2_id INTEGER REFERENCES users(id),
 mutually boolean
 );
-
---INSERT INTO users (id, email,login,name)
---VALUES (1,'aa@bb.com', '11111login11', 'vladimir');
